Get levels table

Retrieves levels table for a given study and release version. Allows for filtering by variables and tables. Wrapper around get_metadata().

In addition to the main get_levels() function, there are two study-specific variations:

get_levels_abcd(): for the ABCD study.

get_levels_hbcd(): for the HBCD study.

They have the same arguments as the get_levels() function, except that the study argument is set to the respective study by default, and should not be set by the user.

Usage

get_levels(study, release = "latest", vars = NULL, tables = NULL)

get_levels_abcd(...)

get_levels_hbcd(...)

Arguments

study

character. The study name. One of "abcd" or "hbcd".

release

character. Release version (Default: "latest").

vars

character (vector). Vector with the names of variables to be included.

tables

character (vector). Vector with the names of tables to be included.

...

Additional arguments passed to the underlying get_levels() function.

Value

Data frame with the levels table.

Examples

get_levels("abcd")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

get_levels("hbcd", release = "1.0")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

get_levels("abcd", vars = c("ab_g_dyn__visit_type"))
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

get_levels("abcd", tables = "ab_g_dyn")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

get_levels_abcd(release = "6.0")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

get_levels_hbcd()
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))