Get DEAP ontology URL

Retrieves the URL for a position in the DEAP ontology tree. Allows to specify a hierarchy order (including removing hierarchy levels) and a single or multiple end node(s) that will show up as selected (and filter the data dictionary table) when the URL is opened.

Usage

get_url_deap(
  nodes,
  order,
  study,
  encode = TRUE,
  as_link = TRUE,
  link_type = "md",
  link_name = NULL
)

Arguments

nodes

list of character vectors (length 2). A list of parent node & end node pairs for each of the nodes that should be selected.

order

character vector. The order of hierarchy levels to display in the ontology tree from the following list:

"study"

"domain"

"sub_domain" (only for ABCD)

"source"

"metric" (only for ABCD)

"atlas" (only for ABCD)

study

character. The name of the NBDC study (one of "abcd" or "hbcd").

encode

logical. Whether or not to encode the URL (Default: TRUE). If FALSE, argument as_link must also be FALSE.

as_link

logical. Whether or not to return the URL formatted as a hyperlink (Default: TRUE).

link_type

character. The type of hyperlink to return (one of "md", the default, or "html").

link_name

character. The name to use for the hyperlink (Default: NULL, i.e., the URL is used as name)

Value

character. The URL or hyperlink for the selected node(s) and study.

Examples

NBDCtools:::get_url_deap(
  nodes = list(c("Standard Variables", "General")),
  order = c("study", "domain", "sub_domain", "source", "metric", "atlas"),
  study = "abcd",
  link_name = "ABCD 'Standard Variables' (standard order; Markdown link)"
)
#> Error in get(data_name, envir = asNamespace("NBDCtools")): object 'lst_dds' not found
NBDCtools:::get_url_deap(
  nodes = list(
    c("Imaging", "Desikan"),
    c("Imaging", "Destrieux")
  ),
  order = c("study", "domain", "atlas", "sub_domain", "metric"),
  study = "abcd",
  link_type = "html",
  link_name = "Desikan & Destrieux atlases (custom order; HTML link)"
)
#> Error in get(data_name, envir = asNamespace("NBDCtools")): object 'lst_dds' not found
NBDCtools:::get_url_deap(
  nodes = list(c("BioSpecimens", "Biological Mother")),
  order = c("study", "domain", "source"),
  study = "hbcd",
  as_link = FALSE # just return the URL, not a link
)
#> Error in get(data_name, envir = asNamespace("NBDCtools")): object 'lst_dds' not found