Get DEAP ontology URL for a specific table

Wrapper around get_url_deap() to get the URL to open a specific table in the DEAP ontology tree.

In addition to the main get_url_deap_table() function, there are two study-specific variations:

get_url_deap_table_abcd(): for the ABCD study.

get_url_deap_table_hbcd(): for the HBCD study.

They have the same arguments as the get_url_deap_table() function, except that the study argument is set to the respective study by default, and should not be set by the user.

Usage

get_url_deap_table(
  table,
  study,
  as_link = TRUE,
  link_type = "md",
  link_name = "name"
)

get_url_deap_table_abcd(...)

get_url_deap_table_hbcd(...)

Arguments

table

character. The table name (table_name in the data dictionary).

study

character. The name of the NBDC study (one of "abcd" or "hbcd").

as_link

logical. Whether or not to return the URL formatted as a hyperlink (Default: TRUE).

link_type

character. The type of hyperlink to return (one of "md", the default, or "html").

link_name

character. The name to use for the hyperlink (one of "name", i.e., use the table name, or "label", i.e., use the table label).

...

Additional arguments passed to the underlying get_url_deap_table() function.

Value

character. The URL or hyperlink for the specified table.

Examples

NBDCtools:::get_url_deap_table("ab_g_dyn", "abcd")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

NBDCtools:::get_url_deap_table("par_visit_data", "hbcd", link_name = "label")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

NBDCtools:::get_url_deap_table_abcd("ab_g_dyn")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))

NBDCtools:::get_url_deap_table_hbcd("par_visit_data", link_name = "label")
#> Error in check_pkgs(pkgs = list(NBDCtoolsData = list(remote = "nbdc-datahub/NBDCtoolsData")),     abort_msg = "NBDCtoolsData package is not installed", quiet = quiet): NBDCtoolsData package is not installed
#> ✖ `NBDCtoolsData`
#> ℹ Please install the missing packages with:
#> 
#> renv::install(c("nbdc-datahub/NBDCtoolsData"))